rename ChoiceAnswer to Assessable
