+++ src/components/documents/Assessable/Hints/index.tsx
