+++ src/components/documents/Assessable/Quiz/styles.module.scss
