+++ src/components/documents/Assessable/_vars.scss
