+++ src/components/documents/Assessable/helpers/scoring.tsx
